requirements.txt
streamlit>=1.36
openai>=1.35
python-dotenv>=1.0
tiktoken>=0.7
pyarrow<18